# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "api",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/v1/projects": {
      "get": {
        "tags": [
          "routes::projects"
        ],
        "operationId": "list_projects",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "routes::projects"
        ],
        "operationId": "create_project",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/projects/{project_key}": {
      "get": {
        "tags": [
          "routes::projects"
        ],
        "operationId": "get_project",
        "parameters": [
          {
            "name": "project_key",
            "in": "path",
            "description": "Project key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "routes::projects"
        ],
        "operationId": "delete_project",
        "parameters": [
          {
            "name": "project_key",
            "in": "path",
            "description": "Project key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        }
      },
      "patch": {
        "tags": [
          "routes::projects"
        ],
        "operationId": "update_project",
        "parameters": [
          {
            "name": "project_key",
            "in": "path",
            "description": "Project key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateProjectRequest": {
        "type": "object",
        "required": [
          "key",
          "name"
        ],
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "key": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "ProjectListResponse": {
        "type": "object",
        "required": [
          "projects"
        ],
        "properties": {
          "projects": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProjectResponse"
            }
          }
        }
      },
...
      "ProjectResponse": {
        "type": "object",
        "required": [
          "id",
          "key",
          "name",
          "owner_id",
          "todo_count",
          "in_progress_count",
          "done_count"
        ],
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "done_count": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "id": {
            "type": "string"
          },
          "in_progress_count": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "key": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "owner_id": {
            "type": "string"
          },
          "todo_count": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
...
      "UpdateProjectRequest": {
        "type": "object",
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
    }
  }
}